feat(scope): allow injected languages to be parsed by treesitter
Adding a config flag to allow injected languages to be parsed by
treesitter. This is crucial for languages like vue, which contain
typescript or javascript in a script tag. When get_node is not called
with ignore_injections = false, the script tag is treated as a single
block, which is not helpful, as it contains most code.