fix(filter): Fix filtering condition check
Check the boolean field of the filter path only when there is a direct
match. Otherwise a filter's path boolean value that is true dominates
the results.

Example:

```lua
opts = {
  sources = {
    recent =  {
      filter = {
        paths = {
          [vim.fn.stdpath("data")] = true,
        }
      }
    }
  }
}
```

This would cause the recent picker to only include paths from the
stdpath("data"). Since the previous condition could never allow any
other path to return true.

But I merely want to enable that path to be included in the results.