feat(explorer): add `open_all`
Add a way to open all directories in the explorer at once, as a
counterpart to the existing `close_all`.

This makes the explorer way more useful to me, making it possible to get
an overview of the files in a project without having to press too many
keys.

This implementation doesn't work recursively, which I guess can be viewed
as either a bug or a feature.