feat(gitbrowse): use upstream branch name when available. See #2613
When branch is not overridden via options, the local name of the current
branch is transformed into the name of the remote branch that
corresponds to the local branch via (transitive) upstream relation (as
established via git branch --set-upstream-to).

When branch is overridden, its upstream branch name is not looked up, as
we are assuming that a user-specified branch name is in the context of
the remote.